refactor(permissions.py): update custom permission IsAuthorOrReadOnly
According to the task requirements, authenticated users are allowed to modify and delete their own content; in all other cases, access is granted for reading only. For convenience, it was decided to modify the permission and add an authentication check for the user. The permission is planned to be used at the project level.